Yeah, none of this form stuff is necessary upon further reflection.
