+++ src/Form/AddDatastreamForm.php
@@ -2,12 +2,11 @@
 
 namespace Drupal\islandora\Form;
 
+use Drupal\Core\Form\FormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Entity\EntityStorageInterface;
 
 use Drupal\Component\Utility\Html;
-
-use Drupal\islandora\Form\Abstracts\ObjectFormBase;
 
 use Symfony\Component\DependencyInjection\ContainerInterface;
 
@@ -18,7 +17,7 @@ use AbstractObject;
  *
  * @package \Drupal\islandora\Form
  */
-class AddDatastreamForm extends ObjectFormBase {
+class AddDatastreamForm extends FormBase {
   /**
    * File entity storage.
    *
@@ -63,7 +62,7 @@ class AddDatastreamForm extends ObjectFormBase {
     $unused_datastreams = "'" . implode("', '", $unused_datastreams) . "'";
     $upload_size = min((int) ini_get('post_max_size'), (int) ini_get('upload_max_filesize'));
 
-    $form = [
+    return [
       '#attributes' => [
         'enctype' => 'multipart/form-data',
       ],
@@ -129,8 +128,6 @@ class AddDatastreamForm extends ObjectFormBase {
         ],
       ],
     ];
-
-    return parent::buildForm($form, $form_state, $object);
   }
 
   /**
