+++ src/Form/DeleteDatastreamVersionForm.php
@@ -2,10 +2,9 @@
 
 namespace Drupal\islandora\Form;
 
+use Drupal\Core\Form\ConfirmFormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Url;
-
-use Drupal\islandora\Form\Abstracts\DatastreamVersionConfirmFormBase;
 
 use AbstractDatastream;
 
@@ -14,7 +13,7 @@ use AbstractDatastream;
  *
  * @package \Drupal\islandora\Form
  */
-class DeleteDatastreamVersionForm extends DatastreamVersionConfirmFormBase {
+class DeleteDatastreamVersionForm extends ConfirmFormBase {
   protected $datastream;
 
   /**
@@ -57,11 +56,8 @@ class DeleteDatastreamVersionForm extends DatastreamVersionConfirmFormBase {
    * {@inheritdoc}
    */
   public function buildForm(array $form, FormStateInterface $form_state, AbstractDatastream $datastream = NULL, $version = NULL) {
-    if (!isset($datastream[$version])) {
-      throw new Exception($this->t('This indicated version could not be found.'));
-    }
-    elseif (count($datastream) < 2) {
-      throw new Exception($this->t('There must be at least two versions in order to delete one.'));
+    if (!isset($datastream[$version]) || count($datastream) < 2) {
+      return drupal_not_found();
     }
 
     $this->datastream = $datastream;
@@ -69,7 +65,7 @@ class DeleteDatastreamVersionForm extends DatastreamVersionConfirmFormBase {
     $form_state->set('datastream', $datastream);
     $form_state->set('version', $version);
 
-    return parent::buildForm($form, $form_state, $datastream, $version);
+    return parent::buildForm($form, $form_state);
   }
 
   /**
