+++ src/Form/ModuleHandlerAdminForm.php
@@ -2,11 +2,36 @@
 
 namespace Drupal\islandora\Form;
 
-use Drupal\islandora\Form\Abstracts\ModuleHandlerAdminForm as AbstractParent;
+use Drupal\Core\Config\ConfigFactoryInterface;
+use Drupal\Core\Form\ConfigFormBase;
+use Drupal\Core\Extension\ModuleHandlerInterface;
+use Symfony\Component\DependencyInjection\ContainerInterface;
 
 /**
- * Helper; poor-man alias of the other class here.
- *
- * @deprecated
+ * Configuration for the Islandora module.
  */
-abstract class ModuleHandlerAdminForm extends AbstractParent {}
+abstract class ModuleHandlerAdminForm extends ConfigFormBase {
+
+  protected $moduleHandler;
+
+  /**
+   * Class constructor.
+   */
+  public function __construct(ConfigFactoryInterface $config_factory, ModuleHandlerInterface $module_handler) {
+    parent::__construct($config_factory);
+    $this->moduleHandler = $module_handler;
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function create(ContainerInterface $container) {
+    // Instantiates this form class.
+    return new static(
+    // Load the service required to construct this class.
+      $container->get('config.factory'),
+      $container->get('module_handler')
+    );
+  }
+
+}
