+++ src/Form/RegenerateObjectDerivativesForm.php
@@ -2,18 +2,18 @@
 
 namespace Drupal\islandora\Form;
 
+use Drupal\Core\Form\ConfirmFormBase;
 use Drupal\Core\Form\FormStateInterface;
 use Drupal\Core\Url;
 
 use AbstractObject;
-use Drupal\islandora\Form\Abstracts\ObjectConfirmFormBase;
 
 /**
  * Derivative regeneration form.
  *
  * @package \Drupal\islandora\Form
  */
-class RegenerateObjectDerivativesForm extends ObjectConfirmFormBase {
+class RegenerateObjectDerivativesForm extends ConfirmFormBase {
 
   /**
    * The object on which we are operating.
@@ -63,7 +63,7 @@ class RegenerateObjectDerivativesForm extends ObjectConfirmFormBase {
   public function buildForm(array $form, FormStateInterface $form_state, AbstractObject $object = NULL) {
     $this->object = $object;
     $form_state->set(['object'], $object->id);
-    return parent::buildForm($form, $form_state, $object);
+    return parent::buildForm($form, $form_state);
   }
 
   /**
